+++ dump_things_service/backends/__init__.py
@@ -142,13 +142,14 @@ class StorageBackend(metaclass=ABCMeta):
     def get_records_of_classes(
         self,
         class_names: Iterable[str],
-        matching: str | None = None,
+        pattern: str | None = None,
     ) -> BackendResultList:
         raise NotImplementedError
 
     @abstractmethod
     def get_all_records(
         self,
+        pattern: str | None = None,
     ) -> BackendResultList:
         raise NotImplementedError