+++ dump_things_service/backends/sqlite.py
@@ -228,13 +228,24 @@ class _SQLiteBackend(StorageBackend):
 
     def get_all_records(
         self,
+        pattern: str | None = None,
     ) -> SQLResultList:
-        statement = select(Thing)
-        for attribute in self.order_by:
-            statement = statement.order_by(Thing.object[attribute]).options(
-                load_only(Thing.id)
+        if pattern is None:
+            statement = text(
+                'select distinct thing.iri, thing.class_name, thing.sort_key, thing.id '
+                'from thing '
+                "ORDER BY thing.sort_key"
             )
-        with Session(self.engine) as session, session.begin():
+        else:
+            statement = text(
+                'select distinct thing.iri, thing.class_name, thing.sort_key, thing.id '
+                'from thing, json_tree(thing.object) '
+                'where lower(json_tree.value) like lower(:pattern) '
+                "and json_tree.type = 'text' ORDER BY thing.sort_key"
+            )
+
+        with self.engine.connect() as connection:
+            rs = connection.execute(statement, parameters={'pattern': pattern})
             return SQLResultList(self.engine).add_info(
                 ResultListInfo(
                     iri=thing.iri,
@@ -242,7 +253,7 @@ class _SQLiteBackend(StorageBackend):
                     sort_key=thing.sort_key,
                     private=thing.id,
                 )
-                for thing in session.scalars(statement).all()
+                for thing in rs
             )